+++ CHANGELOG.md
@@ -17,7 +17,7 @@ Release sections
 
 ## Unreleased
 ### **TODO**
-- Move WiFi functions to seperate library files within ThetisLib
+- Move WiFi functions to separate library files within ThetisLib
 - Implement HTTP API ([#7](https://github.com/Legohead259/Thetis-Firmware/issues/7))
 
 ### Added